doc: component.image mapper jsdocs
